pyproject.toml
[project]
name = "image-mapper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "faiss-cpu>=1.13.2",
    "flask>=3.1.2",
    "numpy>=2.4.2",
    "pandas>=3.0.0",
    "psutil>=7.2.2",
    "scikit-learn>=1.8.0",
    "sentence-transformers>=5.2.2",
]
